build.sh: Add flag to skip unpacking the vendor ramdisk
Some ramdisks may contain privileged mknod commands which
will cause cpio to fail when unpacking the ramdisk when not
run as root, so add a flag and functionality to skip unpacking
such ramdisks and building a boot fs.

Bug: 200194302

+++ build_utils.sh
@@ -277,11 +277,13 @@ function build_boot_images() {
 
     # Remove lib/modules from the vendor ramdisk binary
     # Also execute ${VENDOR_RAMDISK_CMDS} for further modifications
-    ( cd "${MKBOOTIMG_RAMDISK_STAGING_DIR}"
-      cpio -idu --quiet <"${VENDOR_RAMDISK_CPIO}"
-      rm -rf lib/modules
-      eval ${VENDOR_RAMDISK_CMDS}
-    )
+    if [ -z "${SKIP_UNPACKING_RAMDISK}" ]; then
+      ( cd "${MKBOOTIMG_RAMDISK_STAGING_DIR}"
+        cpio -idu --quiet <"${VENDOR_RAMDISK_CPIO}"
+        rm -rf lib/modules
+        eval ${VENDOR_RAMDISK_CMDS}
+      )
+    fi
   fi
 
   if [ -f "${VENDOR_FSTAB}" ]; then
@@ -310,8 +312,12 @@ function build_boot_images() {
 
   if [ "${#MKBOOTIMG_RAMDISK_DIRS[@]}" -gt 0 ]; then
     MKBOOTIMG_RAMDISK_CPIO="${MKBOOTIMG_STAGING_DIR}/ramdisk.cpio"
-    mkbootfs "${MKBOOTIMG_RAMDISK_DIRS[@]}" >"${MKBOOTIMG_RAMDISK_CPIO}"
-    ${RAMDISK_COMPRESS} "${MKBOOTIMG_RAMDISK_CPIO}" >"${DIST_DIR}/ramdisk.${RAMDISK_EXT}"
+    if [ -n "${SKIP_UNPACKING_RAMDISK}" ]; then
+      cat "${VENDOR_RAMDISK_BINARY}" >"${DIST_DIR}/ramdisk.${RAMDISK_EXT}"
+    else
+      mkbootfs "${MKBOOTIMG_RAMDISK_DIRS[@]}" >"${MKBOOTIMG_RAMDISK_CPIO}"
+      ${RAMDISK_COMPRESS} "${MKBOOTIMG_RAMDISK_CPIO}" >"${DIST_DIR}/ramdisk.${RAMDISK_EXT}"
+    fi
   fi
 
   if [ -n "${BUILD_BOOT_IMG}" ]; then
